html/browsers/browsing-the-web/navigating-across-documents/source/navigate-child-function-parent.html

<!doctype html>
<meta charset=utf-8>
<title>Set location from a parent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe></iframe>
<script>
async_test(t => {
  const fr = document.querySelector("iframe")
  fr.contentWindow.location = "resources/dummy.html";
  fr.onload = t.step_func_done(() => {
    assert_equals(fr.contentDocument.referrer, document.URL);
  });
});
</script>
